refactor: make file per event
